Autores y publicaciones

+++ oxl/spiders/bibtec.py
@@ -6,6 +6,7 @@ import urllib2
 import bibtexparser
 import re
 import os
+import os.path
 
 from scrapy.spiders import CrawlSpider, Rule
 from scrapy.selector import HtmlXPathSelector
@@ -28,6 +29,23 @@ class BibtecSpider(CrawlSpider):
         
     }
 
+    def parseauthor(self, url):
+        try:
+            f = urllib.urlopen(url) #Abre la URL del bibtec
+            s = f.read()
+            s = s.split('data-name="')[1]
+            s = s.split('">')[0]
+            #print(s)
+            nombrearchivo = url.split("/")[-1]
+            if os.path.isfile(nombrearchivo+".ttl")==False:
+                f = open(nombrearchivo+".ttl","w+")
+                f.write("@prefix voc: <https://websemantica.icc/rdf/vocabulary#> .\n\n")
+                f.write("<"+url+">\n")
+                f.write('\tvoc:Name "'+s+'" .')
+                f.close()
+        except:
+            "No se pudieron generar los autores"
+
     def parse_item(self, response):
         camposbibtex = ["address","annote","author","booktitle","chapter","crossref","doi","edition","editor","institution","journal","month","number","organitzation","pages","publisher","school","series","title","type","volume","year"]
         url = (str)(response.url)
@@ -39,7 +57,7 @@ class BibtecSpider(CrawlSpider):
             n = len(nbloques)
             for i in range(1,n+1):
                 paths = bloques+'/div['+str(i)+']/div/ul/li'
-                print("Se parsea el bloque "+paths)
+                #print("Se parsea el bloque "+paths)
                 pubs = hxs.select(paths)
                 o  = len(pubs)
                 for j in range(1, o+1):
@@ -51,7 +69,7 @@ class BibtecSpider(CrawlSpider):
                         titulo = titulo.split("<")[0]
                         pathbibtec = paths+'['+str(j)+']/nav/ul/li[2]/div/ul/li[1]/a'
                         pathauthors = paths+'['+str(j)+']/div/span/a'
-                        print(pathauthors)
+                        #print(pathauthors)
                         urlbib = (str)(response.xpath(pathbibtec).extract())
                         urlbib = urlbib.split('"')[1]
                         try:
@@ -65,9 +83,12 @@ class BibtecSpider(CrawlSpider):
                                 if '"' in s3:
                                     s3 = s3.split('"')[1]
                                     authors.append(s3)
+                            setautores = set(authors)
+                            result = list(setautores)
                             autores=""
-                            for k in range(0, len(authors)):
-                                autores += autores+"\n\tvoc:Author <"+authors[k]+"> ;"
+                            for k in range(0, len(result)):
+                                self.parseauthor(result[k])
+                                autores += "\n\tvoc:HasAuthor <"+result[k]+"> ;"
                         except:
                             print ("No hay autores")
 
@@ -82,26 +103,30 @@ class BibtecSpider(CrawlSpider):
                             s = re.sub("\s\s+", " ", s)
                             tipo = s.split("{")[0]
                             tipo = tipo.replace("@","")
-                            print tipo
+                            #print tipo
                             bib_data = parse_string(s,"bibtex")
                             nombrearchivo = urlbib.split("/")[-2]+"-"+urlbib.split("/")[-1]
-                            f = open(nombrearchivo+".ttl","w+")
-                            f.write("@prefix owl: <http://www.w3.org/2002/07/owl#> .\n")
-                            f.write("@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .\n")
-                            f.write("@prefix ns0: <https://websemantica.icc/rdf/vocabulary#> .\n\n")
-                            for entry in bib_data.entries.values():
-                                #print entry.key
-                                #print entry.fields.keys()
-                                f.write("<"+bib_data.entries[entry.key].fields["biburl"]+">")
-                                f.write(autores)
-                                for value in entry.fields.keys():
-                                    #print value ,"=", bib_data.entries[entry.key].fields[value]
-                                    if value in camposbibtex:
-                                        f.write("\n\tvoc:"+value+' "'+bib_data.entries[entry.key].fields[value]+'" ;')
-                            f.seek(-1, os.SEEK_END)
-                            f.truncate()
-                            f.write(".")
-                            f.close()
+                            if os.path.isfile(nombrearchivo+".ttl")==False:
+                                f = open(nombrearchivo+".ttl","w+")
+                                f.write("@prefix owl: <http://www.w3.org/2002/07/owl#> .\n")
+                                f.write("@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .\n")
+                                f.write("@prefix voc: <https://websemantica.icc/rdf/vocabulary#> .\n\n")
+                                for entry in bib_data.entries.values():
+                                    #print entry.key
+                                    #print entry.fields.keys()
+                                    uri = bib_data.entries[entry.key].fields["biburl"]
+                                    uri = uri.replace("bib","html")
+                                    f.write("<"+uri+">")
+                                    f.write("\n\trdf:type voc:"+tipo.capitalize()+" ;")
+                                    f.write(autores)
+                                    for value in entry.fields.keys():
+                                        #print value ,"=", bib_data.entries[entry.key].fields[value]
+                                        if value in camposbibtex:
+                                            f.write("\n\tvoc:"+value.capitalize()+' "'+bib_data.entries[entry.key].fields[value]+'" ;')
+                                f.seek(-1, os.SEEK_END)
+                                f.truncate()
+                                f.write(".")
+                                f.close()
                                 
                         except:
                             print("No se pudo parsear el bibtec")
@@ -115,3 +140,4 @@ class BibtecSpider(CrawlSpider):
 
         except:
             print ("No se pudo parsear "+response.url)
+
